refactor: Put LLMProviders to useLLMProviders hook
